Adds `QueryRunner` class to make querying more declarative
